sentinel-pro/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pyaudio
librosa
numpy
opencv-python
ultralytics
scikit-learn
pyserial
websockets
python-socketio
aiohttp
psycopg2-binary
sqlalchemy
asyncpg
alembic
passlib[bcrypt]
python-dotenv
